Report is generated from the application. Data write throw byte array.
